tackle-api/requirements.txt
Flask==1.1.1
SQLAlchemy==1.3.13
Flask-SQLAlchemy==2.4.1
python-dateutil==2.8.1
flask-cors==3.0.8
pytest==6.2.3
